classcount/python-solution/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0


